misc/HtmlTestSuite/results/test-1289.xhtml

<?xml version="1.0" encoding="UTF-8"?><html xmlns="http://www.w3.org/1999/xhtml"><head><link href="https://www.w3.org/TR/1999/REC-xpath-19991116/#node-sets" rel="help"/>
</head><body>
<script src="/resources/testharness.js"/>
<script src="/resources/testharnessreport.js"/>
<script>
function nodesetToSet(result) {
  const set = new Set();
  for (let node = result.iterateNext(); node; node = result.iterateNext()) {
    set.add(node);
  }
  return set;
}

test(() =&gt; {
  const doc = document.implementation.createHTMLDocument();
  doc.documentElement.innerHTML = '&lt;body&gt;&lt;div&gt;&lt;/div&gt;&lt;/body&gt;';
  const result = nodesetToSet(doc.evaluate('(.//div)[1]|.', doc.documentElement));
  assert_equals(result.size, 2);
  assert_true(result.has(doc.documentElement));
  assert_true(result.has(doc.body.firstChild));
}, '| operator should evaluate both sides of expressions with the same context node');
</script>
</body>
</html>
